Rename `pretext` directory
