+++ packages/doenetml-prototype/src/renderers/pretext-xml/_pass-through-with-tag-and-newline.tsx
