+++ packages/doenetml-prototype/src/renderers/pretext-xml/jsxgraph/listeners.ts
